add code coverage with codecov
fixes https://github.com/zeit/arg/issues/37

I had some difficulties getting codecov to work with the existing
circleci config.  Wasn't really sure why a Docker build was being
done, didn't seem to need to be.  And ... a Docker image was being
built but not actually run?  Not sure what was going on there.

So, I copied from the following circleci config instead:

- https://github.com/zeit/dns-cached-resolve/blob/master/.circleci/config.yml

To enable the codecov bits to work:

- the GH repo will need to be set up in codecov.io
- this will create a token to be used when uploading the codecov file(s)
- that token should be added as an env var in circleci named CODECOV_TOKEN
- voilà, should work