Do not split the arguments if any of them is undefined
The permissive mode will store the undefined arguments
in the `_` array, which is intendented to be processed by another
arguments parsing library, and it may has its own logic to separate
or parse arguments, so we should not process the unknown arguments.

This will change one behavior defined in the test.js.